feat(math): implement under/over parens and shells
Basic functionality has been implemented, likely room for further improvements and refactoring. In the process of implementing this, we have replaced the `reverse` bool used in the function `layout_underoverspreader` with an enum BraceKind, to improve readability. This enum is identical to the existing LineKind used in the function `layout_underoverline`.