WIP
An aside: there should be testing for typst-syntax (as in, testing for AST output).

\# math function call arguments

\## fixed #3802

```typ
\#let ff(x)=x
$ff(f: X → Y)$
```

"unexpected argument: f" raised instead of old "missing argument: x"

fixed as a result of changing/improving how arguments are parsed in math to be more like how they are in the parsing in code.

\## fixed #2052

```typ
\#let func(my-body: none) = my-body

$ func(my-body: a) $
```

named arguments in functions in math can be a normal ident, don't have to be MathIdent like before.

this means it can contain hyphens and start with an underscore (if it isn't just an underscore...) (will also parse single letter paramter names correctly as well).

\## fixing #938

```typ
$mat(..#range(2).map(_ => range(2)))$   //spreading of the argument, so actually works.
$mat(#(..range(2).map(_ => range(2))))$ //passed verbatim, so just reprs the list
```

argument spreading should work in math function calls. previously, the first call above would just repr the list with two dots at the start, and the second call would just repr the list. now the first call works as expected, and the second call's behaviour is the same as before (hasn't changed).

\# attachments

\## fixed #896

```typ
$ sum_(x in ]0, 1]) x $
```

The ( and ) are here to wrap the script content, not to be displayed, so they should not be matched to anything except their closing part.

\## fixed #1435

```typ
a = b_ \
```

Attachments _ and ^ eat newline and align points. Now "empty" math expressions on their own (newline & align points) are forbidden, will get "unexpected ..." error.

\# delimiters

\## fixed #4828

```typ
$1/2(x)$ // parses as 1/2 (x)
$1/10(x)$ // parses as 1/(10(x))
```

also applies to subscripts and superscripts.