Reuse alignment points of previous display math
Closes #1079. Equations with the same label are treated as a single
equation for the purpose of alignment. Performance hasn't been taken
into account just yet. Not sure if the performance hit is substantial,
though there are probably ways to make the current code more performant.
There is also the issue of referencing an equation no longer working, as
multiple elements in the document have the same label. This should be
clarified in the next commit where equation numbering begins to be
implemented.