Add `NumberingMode` to `EquationElem`
Adds the `NumberingMode` struct. Defines and describes the four
numbering modes to be supported. A proper description of this field on
`math.equation` is still missing.

A note for the future: reusing a label for two non-adjacent equations is
problematic. Perhaps this should throw an error (but only when using a
numbering mode where this becomes a problem).