Add labels to math
Adds the `LineLabelElem` in math, and the corresponding syntax node
`MathLineLabel` in `typst-syntax`. We need these to be distinct from the
normal `Label` as the labels in math will only be bound during layout,
unlike normal labels which are bound during evaluation. However, labels
in math have been made to be almost identical to normal labels, making
their difference to end users indistinguishable. They are only "almost"
identical as they disallow underscores in the label text.

Also of note, to mirror that of normal labels, the representation of the
`LineLabelElem` has been changed to be `linebreak()`, matching a
`LinebreakElem`. We do this since labels in math function as linebreaks,
and labels do not exist as content (as they are bound to an element).
This also ensures the current representation of equations is unchanged.

At the moment, all a `LineLabelElem` does when layouted is push a
`MathFragment::Linebreak` to the `MathContext`. Its actual behaviour
will be implemented later, with the equation numbering changes.